backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
anthropic>=0.40.0
facebook-business>=20.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
python-multipart>=0.0.12
python-docx>=1.1.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
cryptography>=43.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
moviepy>=1.0.3
openai-whisper>=20240930
